+++ src/Service/SettingsService.php
@@ -26,12 +26,12 @@ class SettingsService
             $settings->id = 1;
         }
 
-        $settings->title = $liveChatConfig->title;
         $settings->description = $liveChatConfig->description;
         $settings->avatar_asset_id = $liveChatConfig->avatarAssetId;
         $settings->whatsapp_url = $liveChatConfig->whatsAppUrl;
-        $settings->mobile_collapsed = $liveChatConfig->mobileCollapsed;
         $settings->desktop_expanded = $liveChatConfig->desktopExpanded;
+        $settings->show_information_form = $liveChatConfig->showInformationForm;
+        $settings->starter_popup_timer = $liveChatConfig->starterPopupTimer;
         $settings->enabled = $liveChatConfig->enabled;
         $settings->position = $liveChatConfig->position;
         $settings->z_index = $liveChatConfig->zIndex;